Stereotyping of High-Wealth Individuals in Different Ecologies

This dataset originates from a study by Williams et al. investigating whether high-wealth individuals are perceived as having faster life history strategies when they are associated with desperate ecological conditions, compared to hopeful ones.

Usage

data(williams)

Format

A data frame with 112 rows and 25 variables:

subject

Unique subject ID

SexUnres_1 to SexUnres_3

Items measuring perceived sexual unrestrictedness

SexUnres_4_r, SexUnres_5_r

Reverse-coded sexual unrestrictedness items

Impuls_1

Perceived impulsivity item

Impuls_2_r, Impul_3_r

Reverse-coded impulsivity items

Opport_1 to Opport_5

Items measuring opportunity for long-term planning

Opport_6_r

Reverse-coded opportunity item

InvEdu_1_r, InvEdu_2_r

Reverse-coded items measuring perceived educational investment

InvChild_1

Perceived parental investment item

InvChild_2_r

Reverse-coded parental investment item

age

Age of the participant

gender

Gender of the participant (1 = Male, 2 = Female)

ecology

Experimental condition: "Hopeful" or "Desperate"

duration_in_seconds

Duration of the survey in seconds

attention_1, attention_2

Attention check items (1 = correct, 0 = failed)

Source

Data based on materials and methods described by Williams et al.

Details

Participants evaluated a target described as a wealthy individual from either a desperate or hopeful ecology. Measures included perceived sexual unrestrictedness, impulsivity, educational and parental investment, and opportunity for long-term planning.

Examples

data(williams)
str(williams)
#> spc_tbl_ [112 × 25] (S3: spec_tbl_df/tbl_df/tbl/data.frame)
#>  $ subject            : chr [1:112] "A30MP4LXV4MIFD" "A16X5FB3HAFCKN" "A1E9D1OT9VJYDZ" "A16FPOYD7566WI" ...
#>  $ SexUnres_1         : num [1:112] 5 7 2 5 5 5 5 4 6 4 ...
#>  $ SexUnres_2         : num [1:112] 3 7 4 4 5 6 6 6 5 4 ...
#>  $ SexUnres_3         : num [1:112] 2 7 6 5 6 6 5 5 6 4 ...
#>  $ SexUnres_4_r       : num [1:112] 3 4 3 3 3 3 3 2 3 1 ...
#>  $ SexUnres_5_r       : num [1:112] 2 2 3 3 3 2 4 3 2 1 ...
#>  $ Impuls_1           : num [1:112] 3 6 3 4 5 5 5 6 6 1 ...
#>  $ Impuls_2_r         : num [1:112] 3 2 3 4 2 3 3 2 3 1 ...
#>  $ Impul_3_r          : num [1:112] 2 1 3 4 4 3 2 3 2 1 ...
#>  $ Opport_1           : num [1:112] 1 5 3 4 4 4 5 5 5 1 ...
#>  $ Opport_2           : num [1:112] 2 7 5 4 6 3 5 4 4 4 ...
#>  $ Opport_3           : num [1:112] 2 7 5 4 4 4 6 5 5 1 ...
#>  $ Opport_4           : num [1:112] 3 6 3 5 4 6 5 5 4 4 ...
#>  $ Opport_5           : num [1:112] 1 6 3 4 5 4 6 6 4 1 ...
#>  $ Opport_6_r         : num [1:112] 3 2 3 4 3 3 2 2 3 2 ...
#>  $ InvEdu_1_r         : num [1:112] 2 3 2 3 4 3 3 3 3 2 ...
#>  $ InvEdu_2_r         : num [1:112] 3 2 3 4 3 4 2 3 4 1 ...
#>  $ InvChild_1         : num [1:112] 2 5 6 5 5 5 6 5 6 1 ...
#>  $ InvChild_2_r       : num [1:112] 3 2 3 4 3 4 2 4 3 2 ...
#>  $ age                : num [1:112] 34 30 40 35 26 33 33 30 48 33 ...
#>  $ gender             : num [1:112] 1 1 1 2 2 2 2 1 2 2 ...
#>  $ ecology            : chr [1:112] "Hopeful" "Desperate" "Desperate" "Hopeful" ...
#>  $ duration_in_seconds: num [1:112] 164 100 47 31 40 32 98 34 32 87 ...
#>  $ attention_1        : num [1:112] 1 1 1 1 1 1 1 1 1 0 ...
#>  $ attention_2        : num [1:112] 1 1 1 1 1 1 1 1 1 1 ...
#>  - attr(*, "spec")=List of 3
#>   ..$ cols   :List of 25
#>   .. ..$ subject            : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_character" "collector"
#>   .. ..$ SexUnres_1         : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector"
#>   .. ..$ SexUnres_2         : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector"
#>   .. ..$ SexUnres_3         : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector"
#>   .. ..$ SexUnres_4_r       : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector"
#>   .. ..$ SexUnres_5_r       : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector"
#>   .. ..$ Impuls_1           : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector"
#>   .. ..$ Impuls_2_r         : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector"
#>   .. ..$ Impul_3_r          : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector"
#>   .. ..$ Opport_1           : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector"
#>   .. ..$ Opport_2           : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector"
#>   .. ..$ Opport_3           : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector"
#>   .. ..$ Opport_4           : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector"
#>   .. ..$ Opport_5           : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector"
#>   .. ..$ Opport_6_r         : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector"
#>   .. ..$ InvEdu_1_r         : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector"
#>   .. ..$ InvEdu_2_r         : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector"
#>   .. ..$ InvChild_1         : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector"
#>   .. ..$ InvChild_2_r       : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector"
#>   .. ..$ age                : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector"
#>   .. ..$ gender             : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector"
#>   .. ..$ ecology            : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_character" "collector"
#>   .. ..$ duration_in_seconds: list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector"
#>   .. ..$ attention_1        : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector"
#>   .. ..$ attention_2        : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "collector_double" "collector"
#>   ..$ default: list()
#>   .. ..- attr(*, "class")= chr [1:2] "collector_guess" "collector"
#>   ..$ delim  : chr ";"
#>   ..- attr(*, "class")= chr "col_spec"
#>  - attr(*, "problems")=<externalptr> 
table(williams$ecology)
#> 
#> Desperate   Hopeful 
#>        56        56